update-inv-by-gradient
make sure the file can do testing for both spirograph/ cifar
